A pass over the whole tree, and the warnings become errors
Every source parsed by a second front end with the tree's own warning set,
the MSVC analyzer's findings read, clippy over the Rust, and the recent
work re-read. The one that mattered: threads = 0 is one thread in libvpx,
libaom and avm. Each copies it into max_threads and gates every
multithreaded path on max_threads > 1, so VP8, VP9, AV1 by the reference
and AV2 were all decoding on a single core under a comment saying the
library would choose. dav1d's zero genuinely means auto, which is where
the belief came from. All three take the machine's thread count now, from
one shared header.

Also found: codec_mft declared no codecs at all, which made it invisible
to a registry reading declarations, and the module sweep now requires
every codec to declare at least one. verify's SHA-256 ignored a failed
feed and would have vouched for a digest of most of the bytes. codec_mft's
end-of-drain doubt stayed armed across a new run of packets. And matrix
code point 0 -- identity, which is RGB planes -- fell through to BT.709 in
the presenter, and is refused with a sentence instead.

Dead code out: a write-only field in the processed graph, an unread host
pointer in nine modules, a byte-reader accessor nobody called, two test
helpers, dr_wav's implementation compiled twice, and the seven drifted
copies of the test-side module loader, which are one header now.

Then the warnings became errors. /WX and -Werror on mediaperch_flags,
which the root list file links into every target outside external/ by
walking the tree rather than by each target remembering to; the ABI probes
had not. Somebody else's code compiled inside the tree is a target of its
own carrying MEDIAPERCH_EXTERNAL, which the walk skips and nothing else
touches: dr_wav, four files of the VST3 SDK that three targets had each
compiled, the registry calls whose SDK annotations the analyzer objects
to, carved out of sink_asio, and DragonPerch's INI parser. Headers are the
one thing a target boundary cannot separate, so external submodules are
added SYSTEM. Measured: that silences the compiler and not the analyzer,
which reports from inside EbmlBinary.h regardless, so it is told
separately with /analyze:external- -- the one flag here that exists for
somebody else's code.

/Ob2 is replaced with /Ob3 in the base Release flags and the global /Ob3
stays for the submodule that prepends its own /Ob2. Measured: the same
flag twice is silent, and D9025 is a warning /WX does not promote.

The ABI header is compiled as C and as C++, and its C casts were what
-Werror first failed on under Clang. MP_CAST spells a cast in whichever
language is reading, and the clang check no longer waives old-style casts.

Release/AVX2: 314 tests, all passing, under /W4 /WX /analyze. The fuzz
preset builds under -Werror.

+++ modules/dsp/resample/design.cpp
@@ -393,7 +393,8 @@ bool remez_lowpass(std::size_t length, double passband_edge, double stopband_edg
     const std::size_t density = 16;
     const std::size_t points = density * extremals;
     const auto pass_points =
-        std::max<std::size_t>(4, static_cast<std::size_t>(points * pass_width / total));
+        std::max<std::size_t>(4, static_cast<std::size_t>(static_cast<double>(points) *
+                                                          pass_width / total));
     const std::size_t stop_points = std::max<std::size_t>(4, points - pass_points);
 
     std::vector<double> grid;
